implement signature verification and 'nonce' checking
the auth message field will be used as 'nonce'.
rluksd stores this nonce in it's internal stack and drops
all auth messages with double used nonce.
This prevents replay attacks and is an additional feature to
the separation of single shared keys which will also be used
only for one peer.